requirements.txt
fastapi==0.103.2
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
pydantic==1.10.14
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
